70% backend completed

+++ Back_End/src/main/java/com/blinkit/repository/ProductRepository.java
@@ -1,8 +1,8 @@
 package com.blinkit.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import com.blinkit.model.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
 }
